+++ CHANGELOG.md
@@ -113,6 +113,7 @@ All notable changes to this project will be documented in this file.
 - OpenStreetMap example (#1642)
 - Incorrect clipping in TwoColorAreaSeries (#1678)
 - ScreenMin and ScreenMax on Horizontal and Vertical Axes depends on plot bounds (#1652)
+- Windows Forms clipping last line of measured text (#1659)
 
 ## [2.0.0] - 2019-10-19
 ### Added